Use default domain config when probing for everdrive devices, fixes random crashes after resetting on non-X devices
Accessing DOM1 with the faster config in cart_lock seems to cause random
crashes on non-X devices after a simulated reset.
    On a non-x series everdrive:
    1. Get exclusive access to the PI
    2. Make any change to the DOM1 configuration registers
    3. Access DOM1
    4. Change the DOM1 configuration back
    5. Release access to the PI
    6. Use gz's reset command
    7. Play the game for a few minutes
    8. Random crash ???
It's pretty weird.
We need to stay safe and just use the default config when probing in case the
cart isn't an X series device.
This might be an issue on X series devices too, but my test results are
inconclusive.