Update homeboy, patch VC crashes (#85)
This also updates the submodule URL from
https://github.com/krimtonz/homeboy to
https://github.com/PracticeROM/homeboy. To update, run
`git submodule sync && git submodule update`.

The new homeboy version patches out VC crashes, using a fix that
MM Randomizer has been using for a while. When I was using gz
heavily on Wii (basically running it 24/7), I would typically crash
once per day, although it's hard to tell if a crash is a VC crash
or if it's caused by gz instead. I've been testing this fix for a
few days and haven't crashed yet though, which is promising.

The biggest change for gz is that homeboy is now loaded at the end
of MEM1 at 0x817F8000 instead of at the start of MEM2 at 0x90000000.
This makes it easier to add hooks to the emulator, since PowerPC
function calls are limited to a +/- 32MB range, and it will be
necessary for GameCube anyway since there's no MEM2 there. However,
the Wii program loader won't load anything above address 0x80700000,
so instead we initially load homeboy at address 0x80300000 and hack
"_rom_copy_info" in the DOL to have the DOL's C runtime copy it to
0x817F8000 at boot.