Fix scene explorer on Wii VC
Currently the scene explorer on VC is broken because the Z buffer
does not get cleared, meaning the room view clips into the background
instead of being overlaid on top. It almost works to draw a transparent
rectangle with a large "prim depth" instead, but the VC emulator
implementation of gDPSetPrimDepth is buggy and doesn't work for depths
other than 0. I've patched this in homeboy and the scene explorer works
properly now.

Co-authored-by: Tharo <17233964+Thar0@users.noreply.github.com>